+++ test/integration/kubernetes/testdata/invalidarm/arm.json
@@ -3,9 +3,9 @@
     "kind": "Arm",
     "metadata": {
         "name": "arm",
-        "namespace": "arm"
+        "namespace": "invalidarm"
     },
     "spec": {
-        "content": "{\n  \"$schema\": \"https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#\",\n  \"contentVersion\": \"1.0.0.0\",\n  \"functions\": [],\n  \"resources\": [\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Components\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'arm', 'frontend')]\",\n      \"kind\": \"radius.dev/Container@v1alpha1\",\n      \"properties\": {\n        \"run\": {\n          \"container\": {\n            \"image\": 123\n          }\n        },\n        \"uses\": [\n          {\n            \"binding\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'arm', 'backend')).bindings.web]\",\n            \"env\": {\n              \"SERVICE__BACKEND__HOST\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'arm', 'backend')).bindings.web.host]\",\n              \"SERVICE__BACKEND__PORT\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'arm', 'backend')).bindings.web.port]\"\n            }\n          }\n        ],\n        \"bindings\": {\n          \"web\": {\n            \"kind\": \"http\",\n            \"targetPort\": 80\n          }\n        }\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'arm')]\"\n      ]\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Components\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'arm', 'backend')]\",\n      \"kind\": \"radius.dev/Container@v1alpha1\",\n      \"properties\": {\n        \"run\": {\n          \"container\": {\n            \"image\": \"rynowak/backend:0.5.0-dev\"\n          }\n        },\n        \"bindings\": {\n          \"web\": {\n            \"kind\": \"http\",\n            \"targetPort\": 80\n          }\n        }\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'arm')]\"\n      ]\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}', 'radius', 'arm')]\"\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Deployments\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'arm', 'default')]\",\n      \"properties\": {\n        \"components\": [\n          {\n            \"componentName\": \"backend\"\n          },\n          {\n            \"componentName\": \"frontend\"\n          }\n        ]\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'arm')]\",\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'arm', 'backend')]\",\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'arm', 'frontend')]\"\n      ]\n    }\n  ],\n  \"metadata\": {\n    \"_generator\": {\n      \"name\": \"bicep\",\n      \"version\": \"0.1.4.15113\",\n      \"templateHash\": \"10247172844660993266\"\n    }\n  }\n}"
+        "content": "{\n  \"$schema\": \"https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#\",\n  \"contentVersion\": \"1.0.0.0\",\n  \"functions\": [],\n  \"resources\": [\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Components\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'invalidarm', 'frontend')]\",\n      \"kind\": \"radius.dev/Container@v1alpha1\",\n      \"properties\": {\n        \"run\": {\n          \"container\": {\n            \"image\": 123\n          }\n        },\n        \"uses\": [\n          {\n            \"binding\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'invalidarm', 'backend')).bindings.web]\",\n            \"env\": {\n              \"SERVICE__BACKEND__HOST\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'invalidarm', 'backend')).bindings.web.host]\",\n              \"SERVICE__BACKEND__PORT\": \"[[reference(resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'invalidarm', 'backend')).bindings.web.port]\"\n            }\n          }\n        ],\n        \"bindings\": {\n          \"web\": {\n            \"kind\": \"http\",\n            \"targetPort\": 80\n          }\n        }\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'invalidarm')]\"\n      ]\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Components\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'invalidarm', 'backend')]\",\n      \"kind\": \"radius.dev/Container@v1alpha1\",\n      \"properties\": {\n        \"run\": {\n          \"container\": {\n            \"image\": \"rynowak/backend:0.5.0-dev\"\n          }\n        },\n        \"bindings\": {\n          \"web\": {\n            \"kind\": \"http\",\n            \"targetPort\": 80\n          }\n        }\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'invalidarm')]\"\n      ]\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}', 'radius', 'invalidarm')]\"\n    },\n    {\n      \"type\": \"Microsoft.CustomProviders/resourceProviders/Applications/Deployments\",\n      \"apiVersion\": \"2018-09-01-preview\",\n      \"name\": \"[format('{0}/{1}/{2}', 'radius', 'invalidarm', 'default')]\",\n      \"properties\": {\n        \"components\": [\n          {\n            \"componentName\": \"backend\"\n          },\n          {\n            \"componentName\": \"frontend\"\n          }\n        ]\n      },\n      \"dependsOn\": [\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications', 'radius', 'invalidarm')]\",\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'invalidarm', 'backend')]\",\n        \"[resourceId('Microsoft.CustomProviders/resourceProviders/Applications/Components', 'radius', 'invalidarm', 'frontend')]\"\n      ]\n    }\n  ],\n  \"metadata\": {\n    \"_generator\": {\n      \"name\": \"bicep\",\n      \"version\": \"0.1.4.15113\",\n      \"templateHash\": \"10247172844660993266\"\n    }\n  }\n}"
     }
 }
